requirements.txt
streamlit
pandas
pymongo
numpy
smartsheet-python-sdk
firebase-admin